requirements.txt
# Mac软件版本追踪器依赖包

# 核心依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
pandas>=2.1.0
openpyxl>=3.1.0

# 异步支持
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# 数据处理
pydantic>=2.5.0
python-dateutil>=2.8.0
pytz>=2023.3

# 缓存
redis>=5.0.0
diskcache>=5.6.0

# 日志和监控
structlog>=23.2.0
prometheus-client>=0.19.0

# API框架
flask>=3.0.0
flask-restx>=1.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# 任务调度
schedule>=1.2.0
celery>=5.3.0

# 配置管理
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 数据库（可选）
sqlalchemy>=2.0.0
alembic>=1.13.0

# 通知服务
slack-sdk>=3.26.0
requests-oauthlib>=1.3.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# 开发工具
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# 浏览器驱动管理
webdriver-manager>=4.0.0

# 文件处理
xlrd>=2.0.0
xlwt>=1.3.0

# 网络工具
urllib3>=2.1.0
certifi>=2023.11.17

# 数据验证
validators>=0.22.0
jsonschema>=4.20.0

# 加密和安全
cryptography>=41.0.0
bcrypt>=4.1.0

# 性能优化
cachetools>=5.3.0
memory-profiler>=0.61.0

# 图像处理（用于验证码识别，可选）
pillow>=10.1.0
opencv-python>=4.8.0

# 机器学习（用于智能策略选择，可选）
scikit-learn>=1.3.0
numpy>=1.25.0

# 文档生成
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# 部署相关
docker>=6.1.0
kubernetes>=28.1.0

# 监控和告警
sentry-sdk>=1.38.0
newrelic>=9.2.0